Changes as per review

+++ include/ros2_control_abb_driver/abb_system_position_only.hpp
@@ -67,7 +67,7 @@ public:
 
 private:
   // Hardware parameters
-  double robotstudio_port_;
+  size_t robotstudio_port_;
   
   // Store the command for the simulated robot
   std::vector<double> hw_commands_, hw_states_;
